Added in Data
Added in data tables, as described by our Phase l diagram.
Note that the primary keys are just listed as ID's (i.e. bundle_id in the bundle table is just id instead of bundle_id)
Also added a commented-out user model under users/models.py for different users.  This is the basic form of what I was working with for differentiating between vendors/customers, but it and anything built on it is currently broken, hence why it is commented.