New posts now correctly create a product or service too
Creating a new product or service add the product or service to the database and creates a post by that user with the relevant information. Essentially, it should work as before with the Post functionality, but now also creates products and services and bundles and such on the back end too.